ml_utils/requirements.txt
pandas
openpyxl
opencv-python
Pillow
matplotlib
scipy
torch